refactor: move content edition logic into a mixin
